+++ SnowRunner-Tool/ToastNote.cs
@@ -1,0 +1,37 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using Windows.UI.Notifications;
+using Windows.Data.Xml.Dom;
+using System.IO;
+
+namespace SnowRunner_Tool
+{
+    class ToastNote
+    {
+        public static void Notify(string title, string text)
+        {
+            // Get a toast XML template
+            XmlDocument toastXml = ToastNotificationManager.GetTemplateContent(ToastTemplateType.ToastImageAndText03);
+
+            // Fill in the text elements
+            XmlNodeList stringElements = toastXml.GetElementsByTagName("text");
+            stringElements[0].AppendChild(toastXml.CreateTextNode(title));
+            stringElements[1].AppendChild(toastXml.CreateTextNode(text));
+
+            // Specify the absolute path to an image
+            String imagePath = "file:///" + Path.GetFullPath(@"\images\baseline_restore_white_48dp.png");
+            XmlNodeList imageElements = toastXml.GetElementsByTagName("image");
+            imageElements[0].Attributes.GetNamedItem("src").NodeValue = imagePath;
+
+            // Create the toast and attach event listeners
+            ToastNotification toast = new ToastNotification(toastXml);
+
+            // Show the toast. Be sure to specify the AppUserModelId on your application's shortcut!
+            ToastNotificationManager.CreateToastNotifier("SnowRunner-Tool").Show(toast);
+        }
+    }
+}