+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ## v1.0.3.2 (30.5.2020)
 - Add icons to all menu items
+- Refresh table and title after automatic backup
+- Check for update at start, send toast notification
+
 
 ## v1.0.3.1 (24.05.2020)
 - Reveal backup in explorer (right click)